Fix AddMealScreen compile errors: KeyboardOptions import and gram type

+++ app/src/main/java/com/lokma/app/ui/screens/AddMealScreen.kt
@@ -7,6 +7,7 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Check
 import androidx.compose.material3.AlertDialog
@@ -27,7 +28,6 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.text.input.KeyboardOptions
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -144,7 +144,7 @@ fun AddMealScreen() {
                         if (grams != null && validationError == null) {
                             vm.addMeal(foodId = selectedFood.id, grams = grams.toFloat())
                             recentlyAddedFoodId = selectedFood.id
-                            recentlyAddedGrams = grams
+                            recentlyAddedGrams = grams.toFloat()
                             selectedFoodForAdd = null
                         }
                     },
